- Fixed unsoundness caused by wrongsignedness of sizeof:   sizeof is unsigned while VeriFast treated it as signed.   When assigning a sizeof expression to a signed int,   C does an implicit cast, which VeriFast did not see.   This was exploitable (see sizeof_signedness.c). - Typechecking of ghost integral types is now more flexible:   you do not have to type casts anymore to "cast" a ghost integral type   to another ghost integral type. Note that all ghost integegral   types had (and have) the same set of possible values, so the   cast did (and does) nothing.   Because of this typechecker flexibility, not all annotations   containing sizeof had to add a cast.
Fixes #1660.

Credits: how to implement more flexible typechecker is Bart's idea.

+++ examples/arraylist.c
@@ -57,7 +57,7 @@ void list_add(struct arraylist *a, void *v)
     void** newData = malloc((capacity + 100) * sizeof(void*));
     if(newData == 0) abort();
     //@ pointers_split(newData, size);
-    memcpy(newData, data, size * sizeof(void*));
+    memcpy(newData, data, (unsigned int) size * sizeof(void*));
     //@ chars_to_pointers(newData, size);
     a->data = newData;
     a->capacity = capacity + 100;
@@ -78,7 +78,7 @@ void list_remove_nth(struct arraylist *a, int n)
   int size = a->size;
   //@ pointers_split(data, n);
   //@ open pointers(data + n, _, _);
-  memmove(data + n, data + n + 1, (size - n - 1) * sizeof(void *));
+  memmove(data + n, data + n + 1, (unsigned int) (size - n - 1) * sizeof(void *));
   //@ chars_to_pointers(data + n, size - n - 1);
   a->size = a->size - 1;
 }
